refactor(metrics): calculateCodeMetrics を分解など
