+++ src/usecase/generateTsg.ts
@@ -5,10 +5,7 @@ import { abstraction } from '../feature/graph/abstraction';
 import { createGraph } from '../feature/graph/createGraph';
 import { filterGraph } from '../feature/graph/filterGraph';
 import { highlight } from '../feature/graph/highlight';
-import {
-  CodeMetrics,
-  calculateCodeMetrics,
-} from '../feature/metric/calculateCodeMetrics';
+import { calculateCodeMetrics } from '../feature/metric/calculateCodeMetrics';
 import { Graph } from '../feature/graph/models';
 import { writeMetrics } from '../feature/metric/writeMetricsTable';
 import { OptionValues } from '../setting/model';
@@ -18,6 +15,7 @@ import {
   measureInstability,
   writeCouplingData,
 } from '../feature/graph/instability';
+import { CodeMetrics } from '../feature/metric/convertRawToCodeMetrics';
 
 export async function generateTsg(
   commandOptions: OptionValues & { executedScript: string },
